Force string length to be a long, not int.

+++ src/main/java/org/dynjs/runtime/builtins/types/string/DynString.java
@@ -24,7 +24,7 @@ public class DynString extends PrimitiveDynObject {
         super.setPrimitiveValue(value);
         defineOwnProperty(null, "length", new PropertyDescriptor() {
             {
-                set("Value", ((String) value).length());
+                set("Value", (long) ((String) value).length());
                 set("Writable", false);
                 set("Configurable", false);
                 set("Enumerable", false);
